requirements.txt
Django
-e git://github.com/catalpainternational/simple_locations#egg=simple_locations
django-modeltranslation==0.11
